make user id int
git-svn-id: svn://localhost/swayam@120 c64e1095-0faf-4994-8903-70d657f34d2f

+++ prospects/dms/DmsWeb/trunk/src/main/java/com/swayam/dms/web/dao/UserDao.java
@@ -15,7 +15,7 @@ import com.swayam.dms.web.model.User;
  * 
  * @author <a href="mailto:matt@raibledesigns.com">Matt Raible</a>
  */
-public interface UserDao extends GenericDao<User, Long> {
+public interface UserDao extends GenericDao<User, Integer> {
 
     /**
      * Gets users information based on login name.
